[refactor] Update import paths in LoginScreen and RegisterScreen to use shared components
