+++ src/app/core/pages/login/LoginScreen.tsx
@@ -5,15 +5,15 @@ import {
   FormLabel,
   FormControl,
   FormMessage,
-} from "../../../../components/ui/form";
-import { Input } from "../../../../components/ui/input";
-import { Button } from "../../../../components/ui/button";
+} from "../../../shared/components/ui/form";
+import { Input } from "../../../shared/components/ui/input";
+import { Button } from "../../../shared/components/ui/button";
 import {
   Card,
   CardContent,
   CardHeader,
   CardTitle,
-} from "../../../../components/ui/card";
+} from "../../../shared/components/ui/card";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 import { zodResolver } from "@hookform/resolvers/zod";
